v0.1.7
Set up the first project inpage and linked to the homepage via the project
card. All looks to be working, so happy with this.

FOr the project case study, I anticipate adding video content to make the
page a bit more dynamic. Right now, its not showing its depth.

More to come here.